introduce a reference device annotation to group composable previews
