+++ ui/settings/src/main/kotlin/photos/network/ui/settings/composable/ServerSetupItem.kt
@@ -1,0 +1,111 @@
+package photos.network.ui.settings.composable
+
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.core.tween
+import androidx.compose.animation.expandVertically
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.shrinkVertically
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.KeyboardArrowDown
+import androidx.compose.material.icons.filled.KeyboardArrowRight
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import photos.network.api.ServerStatus
+import photos.network.ui.settings.FormInput
+import photos.network.ui.settings.R
+
+
+@Composable
+fun ServerSetupItem(
+    modifier: Modifier = Modifier,
+    isExpanded: Boolean = false,
+    serverHost: String = "",
+    onServerHostUpdated: (String) -> Unit = {},
+    isHostVerified: Boolean = false,
+    clientId: String = "",
+    onClientIdUpdated: (String) -> Unit = {},
+    isClientIdVerified: Boolean = false,
+    serverStatus: ServerStatus = ServerStatus.UNAVAILABLE,
+    onServerSetupClicked: () -> Unit = {},
+) {
+    val serverSetupLabel = if (serverStatus != ServerStatus.AVAILABLE) {
+        stringResource(id = R.string.settings_item_server_setup)
+    } else {
+        stringResource(id = R.string.settings_item_server_update)
+    }
+    Surface(
+        modifier = modifier
+            .clickable(
+                onClickLabel = serverSetupLabel,
+            ) {
+                onServerSetupClicked()
+            },
+    ) {
+        Row(
+            modifier = Modifier
+                .padding(16.dp),
+        ) {
+            Text(
+                modifier = Modifier.weight(1f),
+                text = serverSetupLabel,
+            )
+            if (isExpanded) {
+                Icon(
+                    imageVector = Icons.Default.KeyboardArrowDown,
+                    contentDescription = null,
+                )
+            } else {
+                Icon(
+                    imageVector = Icons.Default.KeyboardArrowRight,
+                    contentDescription = null,
+                )
+            }
+        }
+    }
+
+    Column {
+        // server host
+        AnimatedVisibility(
+            visible = isExpanded,
+            enter = fadeIn(animationSpec = tween(durationMillis = 1000)) + expandVertically(),
+            exit = shrinkVertically(animationSpec = tween(durationMillis = 500, delayMillis = 0)),
+        ) {
+            FormInput(
+                modifier = modifier,
+                label = "Host",
+                value = serverHost,
+                hint = "https://",
+                onValueChanged = {
+                    onServerHostUpdated(it)
+                },
+                showTrailingIcon = isHostVerified,
+            )
+        }
+
+        // client id
+        AnimatedVisibility(
+            visible = isExpanded && isHostVerified,
+            enter = fadeIn(animationSpec = tween(durationMillis = 1000)) + expandVertically(),
+            exit = shrinkVertically(animationSpec = tween(durationMillis = 500, delayMillis = 0)),
+        ) {
+            FormInput(
+                modifier = modifier,
+                label = "Client ID",
+                value = clientId,
+                onValueChanged = {
+                    onClientIdUpdated(it)
+                },
+                showTrailingIcon = isClientIdVerified,
+            )
+        }
+    }
+}
